Consolidate header files
